requirements.txt
appnope==0.1.4
asttokens==3.0.0
beautifulsoup4==4.13.3
blinker>=1.9.0
bs4==0.0.1
cachelib==0.13.0
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.13
decorator==5.2.1
executing==2.2.1
Flask==3.1.2
Flask-Caching==2.3.1
Flask-MySQLdb==2.0.0
Flask-WTF==1.1.1
fonttools==4.39.4
gunicorn==23.0.0
ipykernel==6.29.5
ipython==8.16.1
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.3.0
kiwisolver==1.4.8
loader==2017.9.11
MarkupSafe==2.1.3
matplotlib==3.10.1
matplotlib-inline==0.1.7
mysqlclient==2.2.7
narwhals==1.33.0
nest-asyncio==1.6.0
numpy==1.26.4
packaging==25.0
pandas==2.3.0
parso==0.8.4
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
plotly==6.0.1
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==26.3.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
soupsieve==2.6
stack-data==0.6.3
threadpoolctl==3.6.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.0
tzdata==2025.2
wcwidth==0.2.13
Werkzeug==3.1.3
WTForms==3.2.1